Introducing Custom Exception and Fluent Validations

+++ HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using HR.LeaveManagement.Domain;
@@ -26,8 +27,9 @@ namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
 
-            //if (leaveType == null)
-            //    throw new NotFoundException(nameof(LeaveType), request.Id);
+            //notfound exception takes two params as key value pairs e.g name that was passed and the key id passed with it
+            if (leaveType == null)
+                throw new NotFoundException(nameof(LeaveType), request.Id);
 
             await _leaveTypeRepository.Delete(leaveType);
